Reorganize the org.scijava.io classes
This moves Location stuff to org.scijava.io.location,
and DataHandle stuff to org.scijava.io.handle.

It breaks backwards compatibility, but the Location API was not being
used by any SJC-based systems yet. We make the change now to minimize
future damage.
