+++ src/main/java/org/scijava/io/handle/AbstractDataHandle.java
@@ -29,10 +29,11 @@
  * #L%
  */
 
-package org.scijava.io;
+package org.scijava.io.handle;
 
 import java.nio.ByteOrder;
 
+import org.scijava.io.location.Location;
 import org.scijava.plugin.AbstractWrapperPlugin;
 
 /**
