+++ src/main/java/org/scijava/io/location/AbstractLocation.java
@@ -29,35 +29,13 @@
  * #L%
  */
 
-package org.scijava.io;
-
-import org.scijava.plugin.WrapperService;
-import org.scijava.service.SciJavaService;
+package org.scijava.io.location;
 
 /**
- * Interface for low-level data I/O: reading and writing bytes using
- * {@link DataHandle}s.
+ * Abstract base class for {@link Location} implementations.
  * 
  * @author Curtis Rueden
- * @see IOService
- * @see Location
  */
-public interface DataHandleService extends
-	WrapperService<Location, DataHandle<Location>>, SciJavaService
-{
-
-	// -- PTService methods --
-
-	@Override
-	@SuppressWarnings({ "rawtypes", "unchecked" })
-	default Class<DataHandle<Location>> getPluginType() {
-		return (Class) DataHandle.class;
-	}
-
-	// -- Typed methods --
-
-	@Override
-	default Class<Location> getType() {
-		return Location.class;
-	}
+public abstract class AbstractLocation implements Location {
+	// NB: No implementation needed.
 }
