+++ src/test/java/org/scijava/ContextCreationTest.java
@@ -95,9 +95,9 @@ public class ContextCreationTest {
 				org.scijava.display.DefaultDisplayService.class,
 				org.scijava.event.DefaultEventHistory.class,
 				org.scijava.input.DefaultInputService.class,
-				org.scijava.io.DefaultDataHandleService.class,
 				org.scijava.io.DefaultIOService.class,
 				org.scijava.io.DefaultRecentFileService.class,
+				org.scijava.io.handle.DefaultDataHandleService.class,
 				org.scijava.io.nio.DefaultNIOService.class,
 				org.scijava.main.DefaultMainService.class,
 				org.scijava.menu.DefaultMenuService.class,
